LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-container-dynamic-expected.txt


FAIL Custom Functions: @container responds to changes assert_equals: --actual before resize expected "A" but got "--f()"
